members logic moved to controller and validator
