+++ middlewares/membersValidator.js
@@ -1,0 +1,17 @@
+const { body, validationResult } = require('express-validator');
+
+const membersValidation = [
+    body('name').isString(),
+];
+
+const validationHandler = (req, res, next) =>{ 
+    const errors = validationResult(req);
+    if (errors.isEmpty()) {
+      return next();
+    }
+    return res.status(400).json({ errors: errors.array() });
+};
+
+module.exports = {
+    membersValidation, validationHandler
+}
